==> notebooks/01_drug_dispensed_eda.ipynb <==
%load_ext autoreload
%autoreload 2

import os, sys
from dotenv import load_dotenv
load_dotenv("../.env")  # take environment variables
PROJECT_ROOT = os.environ.get("PROJECT_ROOT")
sys.path.append(PROJECT_ROOT)

import numpy as np
import pandas as pd
from glob import glob

from datetime import datetime
import utils.PATHS as PATHS
import utils.utils as utils
# import utils.emr_utils as emr_utils
# import utils.load_utils as load_utils

## Notes

The following were observed in the exploration below:
* Potentially useful prescription-related columns:
    - `Generic Drug Name`
* Consider datetime columns as reference for how long medication has been going in relation to lipid-lowering treatment

dd_fp_list = glob(os.path.join(PATHS.DRUG_DISPENSED, "*.csv"))

test = pd.read_csv(dd_fp_list[0])
test.info()

## Questions

1. What to do when combinations within the lookback were observed?
2. The statin + ____ prescriptions should be addressed within lookback?

# # SPECIAL CORRECTION FACTORS
# {
#     # "pravastatin sodium": ,  # 20MG
#     "ezetimibe,simvastatin tab" : , # 10/10, 10/20
#     "ezetimibe": , # 10mg
#     "simvastatin": , # 10mg, 20mg
#     "atorvastatin": , # 10mg, 40mg, 20mg
#     "rosuvastatin": , # 10MG, 20MG


#     "lovastatin": , # 20mg
#     "metronidazole, nystatin": , # 500mg, 100,000 unit    
#     "nystatin": , # 100,000IU
# } 

test = test.dropna(subset=["Generic Drug Name"])

test["Generic Drug Name"] = test["Generic Drug Name"].str.lower()

drug_name = "nystatin"
# drug_name = "statin"
sub = test[test["Generic Drug Name"].str.startswith(drug_name)][["Generic Drug Name", "Drug Strength"]]
display(sub.head())
display(sub["Generic Drug Name"].unique())
display(sub["Drug Strength"].unique())

test[["Generic Drug Name", "Prescribed Quantity UOM",]]

test[["Prescribed Quantity UOM", "Item Primary UOM", "Drug Strength", "Drug Form"]].head(20)

test.iloc[:,-8:]

test[["Add Date", "Start Date"]].head(200)

test[["Case Start Date", "Case End Date"]].head(200)

test[["Drug Prescribed Date From", "Drug Prescribed Date To"]].head(200)

test[["Drug Code", "Drug Name", "Drug Strength", "Drug Form", "Generic Drug Id", "Generic Drug Name", "DRG Code", "DRG Code (AR-DRG V6.0)"]]

# "DRG Code", "DRG Code (AR-DRG V6.0)" # diagnosis related group
test["DRG Code"].value_counts()

test["DRG Code (AR-DRG V6.0)"].value_counts()

test["Drug Status"].value_counts()

test.head()

## End.
