==> notebooks/02_general_lab_eda.ipynb <==
%load_ext autoreload
%autoreload 2

import os, sys
from dotenv import load_dotenv
load_dotenv("../.env")  # take environment variables
PROJECT_ROOT = os.environ.get("PROJECT_ROOT")
sys.path.append(PROJECT_ROOT)


import numpy as np
import pandas as pd
from glob import glob
from tqdm import tqdm

import utils.PATHS as PATHS
import utils.utils as utils
# import utils.emr_utils as emr_utils
# import utils.load_utils as load_utils

## Notes

The following were observed in the exploration below:
* Potentially useful columns:
  - `Lab Resulted Order Test Description` is more intuitive in mapping to a the `Result value` column
  - `Lab Test Description` is broader. Such as it may consist of different `Lab Resulted Order Test Description`
* Reliable datetime columns (for Index Date of highest LDL-C), ordered by most suitable:
  - `Specimen Received Date`
  - `Specimen Collection Date`
  - `Admit Date`
* **Issue**: Lab Test Shortext Descriptions provided by collaborators dont match with SGH vocabulary. Resolve with clinician input.

lab_fp_list = glob(os.path.join(PATHS.LABS, "*.csv"))

print(lab_fp_list[0])
test = pd.read_csv(lab_fp_list[0])
test.info()

# lab results columns
lab_cols = [
    "Lab Test Code",
    "Lab Test Description",
    "Lab Test Type",
    "Lab Test Type (KKH Only)",
    "Lab Test Display",
    "Result Value"
]
test[lab_cols]

# ROTD per TD
test[["Lab Test Description", "Lab Resulted Order Test Description", "Result Value"]].groupby("Lab Test Description")["Lab Resulted Order Test Description"].apply(set)

# TD per ROTD : See notes on findings
test[["Lab Test Description", "Lab Resulted Order Test Description", "Result Value"]].groupby("Lab Resulted Order Test Description")["Lab Test Description"].apply(set)

test["Lab Test Description"].value_counts()

other_lab_cols = [
    "Lab Test Code",
    "Lab Resulted Order Test Description",
    "Lab Resulted Order Test Type",    
    "Lab Resulted Order Test Type (KKH Only)",
    "Lab Resulted Order Test Display",
    "Result Value"    
]
test[other_lab_cols]

test["Lab Resulted Order Test Description"].value_counts()

# date columns
dt_cols = [
    "Result Comment Date",
    "Specimen Received Date",
    "Specimen Received Date.1",
    "Admit Date",
    "Specimen Collection Date",
    "Visit Date"
]
test[dt_cols]

# check LDL-C Tests
(test.loc[test["Lab Resulted Order Test Description"].str.contains("LDL", case=False)]
 [["Patient ID", "Lab Resulted Order Test Description", "Result Value"]])["Lab Resulted Order Test Description"].unique()

# check LDL-C Tests
(test.loc[test["Lab Test Description"].str.contains("LDL", case=False)]
 [["Patient ID", "Lab Test Description", "Result Value"]])["Lab Test Description"].unique()

# # check values on secondary conditions related to hypocholesterolemia
# for i in utils.secondary_labs_st:
#     print(i.upper())
# [i.lower() for i in utils.secondary_labs_st]

# test orders dictionary
df_list = [pd.read_csv(path, low_memory=False) for path in tqdm(lab_fp_list)]
df = pd.concat(df_list, ignore_index=True)

# df.info()

# create vocabulary of test orders
test_orders_vocab = df["Lab Resulted Order Test Description"].unique().tolist()

[i for i in test_orders_vocab if "tsh" in i.lower()]

[i for i in test_orders_vocab if "bilirubin" in i.lower()]

[i for i in test_orders_vocab if "protein" in i.lower()]

[i for i in test_orders_vocab if "urine" in i.lower() and "protein" in i.lower()]

[i for i in test_orders_vocab if "creatinine" in i.lower()]

[i for i in test_orders_vocab if "egfr" in i.lower()]

# efgr - Estimated glomerular filtration rate
[i for i in test_orders_vocab if "glomerular" in i.lower()]

[i for i in test_orders_vocab if "hba1c" in i.lower()]

[i for i in test_orders_vocab if "hba" in i.lower()]

[i for i in test_orders_vocab if "glucose" in i.lower()]

[i for i in test_orders_vocab if "rapid" in i.lower()]

[i for i in test_orders_vocab if "fasting" in i.lower()]

[i for i in test_orders_vocab if "poct" in i.lower() and "creatinine" in i.lower()]

[i for i in test_orders_vocab if "poct" in i.lower()]

## End.
