==> notebooks/04_labs_filtering.ipynb <==
# Labs Filtering

This notebook covers the creation of references for the following Stage 1 filtering criteria:
1. `Labs` `List` Age is 18 and over with valid low-density lipoprotein cholesterol (LDL-C) test result
2.  `Labs` `List` Existing triglycerides (TG) measurements and thresholded at 5.645 mmol/L
3. `Labs` `Dictionary` Date of highest LDL-C measurement (Index Date)
5. `Labs` `List` Those with secondary causes as proxied by lab results thresholded using Table 3.

## Move to llt notebook
6. `Drugs` `List` Those with lipid lowering treatments (LLT) within 1 year to 6 weeks of their Index Date
7. `Labs` `Dictionary` `Modified` Create LDL-C measurement dictionary with actual and scaled values (i.e., based on LLT)

## To-do:
4. `Diagnosis` `List` Those with secondary causes within 1 year (+/-) of Index Date

%load_ext autoreload
%autoreload 2

import os, sys
from dotenv import load_dotenv
load_dotenv("../.env")  # take environment variables
PROJECT_ROOT = os.environ.get("PROJECT_ROOT")
sys.path.append(PROJECT_ROOT)

import numpy as np
import pandas as pd
from glob import glob
from tqdm import tqdm
from datetime import datetime

import utils.PATHS as PATHS
import utils.utils as utils
# import utils.emr_utils as emr_utils
# import utils.load_utils as load_utils

## 1. Valid LDL-C and Age >= 18 patient list

### Notes:
* As LDL-C measurements will be the primary constraint to this referral algorithm, this criteria and refernece dataset (i.e., general labs) is the maximum scope
* Patient not in the general labs dataset will not be accounted for by the referral algorithm
* We are using the following as test names for LDL-C as they appear in the dataset:
  - `LDL-CHOLESTEROL,CALCULATED`; Units: MMOL/L
  - `LDL-CHOLESTEROL,DIRECT`; Units: MMOL/L
* Both CALCULATED and DIRECT measurements are handled the same way.
* Few cases with LDL-C result value of '> 10.30', these were pegged to a value of **10.30**--this risks spurious results in selecting the Index Date for these particular cases; we are not handling this issue for now.

lab_fp_list = glob(os.path.join(PATHS.LABS, "*.csv"))

# general labs dataframe
df_list = [pd.read_csv(path, low_memory=False) for path in tqdm(lab_fp_list)]
df = pd.concat(df_list, ignore_index=True)

df.info()

df['Lab Resulted Order Test Display'].unique()

df['Lab Resulted Order Test Type'].unique()

df['Result Comment Date'].unique()

# Total unique patients
df["Patient ID"].unique().size

# test results reference column
pid_col = "Patient ID"
result_col = "Result Value"
test_name_col = "Lab Resulted Order Test Description"
query_str = "LDL" # lipoprotein, 

# check potential LDL-C tests naming
(
    df.loc[df[test_name_col].str.contains(query_str, case=False)]
    [[pid_col, test_name_col, result_col]]
)[test_name_col].unique()

query_str = "lipoprotein"
(
    df.loc[df[test_name_col].str.contains(query_str, case=False)]
    [[pid_col, test_name_col, result_col]]
)[test_name_col].unique()

# most are nto relevant, stick with LDL-CHOLESTEROL,CALCULATED and
# LDL-CHOLESTEROL,DIRECT
query_str = "LD"
(
    df.loc[df[test_name_col].str.contains(query_str, case=False)]
    [[pid_col, test_name_col, result_col]]
)[test_name_col].unique()

def not_startswithdigit(x):
    try:
        return not x[0].isdigit()
    except:
        return False

# full data
nonfloat = df.loc[df[result_col].apply(lambda x: not_startswithdigit(x))][result_col].unique()
nonfloat[0:20]

test_names = [
    'LDL-CHOLESTEROL,CALCULATED',
    'LDL-CHOLESTEROL,DIRECT'
]
ldlc_all = df.loc[df[test_name_col].isin(test_names)]

nonfloat_ldlc = ldlc_all.loc[ldlc_all[result_col].apply(lambda x: not_startswithdigit(x))][result_col].unique()
nonfloat_ldlc

# address '> 10.30' string
ldlc_all.loc[ldlc_all[result_col] == '> 10.30', "Result Value"] = 10.3

# Validity filtering: Exclude ND and NaN values
# drop those with 'ND' results : Invalid
ldlc_valid = ldlc_all.loc[ldlc_all[result_col] != 'ND']

# coerce results to float
# ldlc_valid.loc[:,result_col] = ldlc_valid[result_col].astype(float)
# use this as .loc back-casts to object type (right hand side)
ldlc_valid[result_col] = ldlc_valid[result_col].astype(float) 

# remove nans
ldlc_valid = ldlc_valid.dropna(subset=result_col)

# pre filtering of valid values
display(ldlc_all.shape)

# post filtering
display(ldlc_valid.shape)

# overall summary statistics
ldlc_valid[[test_name_col, result_col]].describe()

# stratified
grouper = ldlc_valid[[test_name_col, result_col]].groupby(test_name_col)
for name, group_data in grouper:
    print(name)
    display(group_data.set_index(test_name_col).describe())

# sanity checks : nans and ambiguities per patient
tqdm.pandas()

# nan DOBs
display(ldlc_valid["Date of Birth"].isna().sum())

# multiple dob per patient (including nan) : NONE, all pids have 1 unique DOB
(ldlc_valid.groupby("Patient ID")["Date of Birth"]
 .progress_apply(set)
 .progress_apply(list)
 .progress_apply(lambda x: len(x))
).sort_values()



# create age col
ref_date = "2024-01-01"
ldlc_valid["Date of Birth"] = pd.to_datetime(ldlc_valid["Date of Birth"])
ldlc_valid["Age"] = ldlc_valid["Date of Birth"].apply(utils.get_age, args=("%Y-%m-%d", ref_date))

ldlc_valid_18 = ldlc_valid.loc[ldlc_valid["Age"] >= 18]

# pre filtering of valid values
display(ldlc_all.shape)

# post filtering
display(ldlc_valid.shape)

# post filtering age
display(ldlc_valid_18.shape)

ldlc_valid_18["Age"]

ldlc_valid_18_plist = ldlc_valid_18[pid_col].unique().tolist()
len(ldlc_valid_18_plist)

## 2. Existing triglycerides (TG) measurements and thresholded at 5.645 mmol/L

NUHS Reference: "We also exclude those patients who have had a triglycerides test with a value greater than or equal to 5.645 mmol/L **more than once.**"

### Notes:
* Count occurence of TG >= 5.645 mmol/L. If more than 1, then exclude.
* Only using 'TRIGLYCERIDES', excludng those with 'TRIGLYCERIDES,FLUID' and 'TRIGLYCERIDES,URINE'; To be verified.
* NO TIME HORIZON for this criteria. All TG measurements are accounted for regardless of date.

# tricglycerides
query_str = "tri"
(
    df.loc[df[test_name_col].str.contains(query_str, case=False)]
    [[pid_col, test_name_col, result_col]]
)[test_name_col].unique()

# prep tg df
tg = df.loc[df[test_name_col].isin(["TRIGLYCERIDES"])]
tg[result_col][tg[result_col].apply(lambda x: not_startswithdigit(x))].unique()

# address string values
tg.loc[tg[result_col] == '< 0.10', "Result Value"] = 0.099 # arbitrary
tg.loc[tg[result_col] == '> 50.00', "Result Value"] = 50.0099 # arbitrary

# Validity filtering: Exclude ND and NaN values
# drop those with 'ND' results : Invalid
tg_cleaned = tg.loc[tg[result_col] != 'ND']

# coerce results to float
# ldlc_valid.loc[:,result_col] = ldlc_valid[result_col].astype(float)
# use this as .loc back-casts to object type (right hand side)
tg_cleaned[result_col] = tg_cleaned[result_col].astype(float) 

# remove nans
tg_cleaned = tg_cleaned.dropna(subset=result_col)

tg_cleaned[result_col].describe()

# create patient list with TG >= 5.645 more than once
# NO TIME HORIZON for this criteria
tg_exclude_plist = []
for pid, p_data in tg_cleaned.groupby(pid_col):
    temp = p_data.drop_duplicates()
    if temp.shape[0] > 1:
        occurence = temp[result_col].apply(lambda x: x >= 5.645).sum()
        if occurence > 1:
            tg_exclude_plist.append(pid)

print(len(ldlc_valid_18_plist))
print(len(tg_exclude_plist))
running_plist = list(set(ldlc_valid_18_plist).difference(set(tg_exclude_plist)))
print("Remaining: ", len(running_plist))

print("Removed: ", len(ldlc_valid_18_plist) - len(running_plist))

## 3. Date of highest LDL-C measurement (Index Date)

This step creates a dictionary of dates and values of highest LDL-C measurement for patients that are included (so far) at this stage.

ldlc_valid_18_nontg = ldlc_valid_18.loc[ldlc_valid_18[pid_col].isin(running_plist)]
display(ldlc_valid_18_nontg[pid_col].unique().size)
# verified, can save

# prep date col
date_col = "Specimen Received Date"
ldlc_valid_18_nontg[date_col] = pd.to_datetime(ldlc_valid_18_nontg[date_col])

ldlc_index_dict = {} # can save this dict
for pid, p_data in ldlc_valid_18_nontg.groupby(pid_col):
    temp = p_data.drop_duplicates().sort_values(by=result_col)
    # index -1 is the max (sort is least to most default)
    max_ldlc = temp.iloc[-1][result_col] # max ldlc val
    index_date = temp.iloc[-1][date_col] # index date
    ldlc_index_dict[pid] = {
        "LDL-C Max" : max_ldlc,
        "Index Date" : index_date,
    }

# either use this or the dict
ldlc_valid_18_nontg_index = pd.DataFrame().from_dict(ldlc_index_dict, orient='index')
ldlc_valid_18_nontg_index.index.name = 'Patient ID'
ldlc_valid_18_nontg_index.to_csv(
    os.path.join(PROJECT_ROOT, "results", "ldlc_valid_18_nontg_index.csv"), 
)

## 4. With lab-based secondary causes

#### Notes:

* Thresholds used:
  - Threshold for `HBA1C` was converted from mmol/L to percentages to suit the dataset units. We used 9% threshold from the Mayo Clinic document instead of the conversion from the NUHS value i.e., 9.6%.
  - Threshold for `TSH RECEPTOR ANTIBODY` and `TSH RECEPTOR ANTIBODIES (TRAb)` were converted from mU/L to IU/L to suit the dataset.
  - The rest of the thresholds were kept as is.
* All test records with `ND` values were removed
* Specific cleaning done:
  - `BILIRUBIN,TOTAL`
    * "< 3" changed to 2.99
    * "< 2" changed to 1.99
  - `URINE TOTAL PROTEIN`
    * "> 20.0" and "> 20" changed to 20.1
    * "< 0.023" changed to 0.0229
  - `CREATININE`
    * "< 15" changed to 14.99
    * "> 2200" changed to 2200.1
  - `HBA1C`
    * "> 20.1" changed to 20.2
    * "< 4.2" changed to 4.19
  - `TSH RECEPTOR ANTIBODY`
    * "> 40.0" changed to 40.1
  - `TSH RECEPTOR ANTIBODIES (TRAb)`
    * "< 0.90" changed to 0.89 # will cause issues with the threshold for 92 patients
    * "< 1.10" changed to 1.09 # will cause issues with the threshold for 92 patients
    * "> 40.00" changed to 40.1


### 4.1 Thresholds

# thresholds dictionary
secondary_thresholds = {
    "TSH RECEPTOR ANTIBODY": (lambda x: x >= 0.01), # tsh,  mU/L (milli-Units per litre); IU/L (CONVERTED)
    "TSH RECEPTOR ANTIBODIES (TRAb)": (lambda x: x >= 0.01), # tsh,  mU/L (milli-Units per litre); IU/L (CONVERTED)\
    
    "BILIRUBIN,TOTAL" : (lambda x: x > 34.208), # micro-mol/L; UMOL/L (KEEP THRESHOLD)
    
    # "PROTEIN,TOTAL" : (lambda x: x > 3), # grams (DISCARD due to dist of values)
    "URINE TOTAL PROTEIN" : (lambda x: x > 3), # grams, G/L; KEEP
    "PROTEIN,24HR URINE" : (lambda x: x > 3), # grams, G/day; KEEP
    
    "CREATININE" : (lambda x: x > 229.84), # micro-mol/L; UMOL/L; KEEP
    
    "HBA1C" : (lambda x: x > 81.861), # milli-mol/L; percentage (change to percentage threshold 9% from Mayo clinic document, 9.6% using conversion from NUHS threshold)
    
    "GLUCOSE FASTING" : (lambda x: x > 12.22), # milli-mol/L; mmol/L; KEEP
    
    "MDRD eGFR" : (lambda x: x < 15), # efgr, <15 ml/min; NO UNIT; KEEP
    "CORRECTED CALC CG eGFR" : (lambda x: x < 15), # efgr, < 15 ml/min; NO UNIT; KEEP
    "CKD-EPI eGFR": (lambda x: x < 15), # efgr, < 15 ml/min; NO UNIT; KEEP
}

### 4.2 Data cleaning

# thresholds dictionary
value_cleaner = {
    "TSH RECEPTOR ANTIBODY": {
        "> 40.0" : 40.1,
    },
    "TSH RECEPTOR ANTIBODIES (TRAb)": {
        "< 0.90" : 0.89, # will cause issues with the threshold for 92 patients
        "< 1.10" : 1.09, # will cause issues with the threshold for 92 patients
        "> 40.00": 40.1,
    },
    "BILIRUBIN,TOTAL" : {
        "< 3": 2.99,
        "< 2": 1.99,
    },
    "URINE TOTAL PROTEIN" : {
        "> 20.0": 20.1,
        "> 20" : 20.1,
        "< 0.023": 0.0229,
    },
    "CREATININE" : {
        "< 15" : 14.99,
        "> 2200" : 2200.1,
    },
    "HBA1C" : {
        "> 20.1" : 20.2,
        "< 4.2" : 4.19,
    },
}

def clean_values(row):
    mapper = value_cleaner.get(row[test_name_col])
    if mapper:
        row[result_col] = mapper.get(row[result_col], row[result_col])
        return row
    else:
        return row

tqdm.pandas()
secondary_labs = df.loc[df[test_name_col].isin(secondary_thresholds.keys())]

# update values
secondary_labs_cleaned = secondary_labs.progress_apply(lambda row: clean_values(row), axis=1)

# sanity check : will show only ND values are left
display(secondary_labs_cleaned[result_col][secondary_labs_cleaned[result_col].apply(lambda x: not_startswithdigit(x))].unique())

# address ND and NaN values; cast to float
secondary_labs_cleaned = secondary_labs_cleaned.loc[secondary_labs_cleaned[result_col] != 'ND']
secondary_labs_cleaned[result_col] = secondary_labs_cleaned[result_col].astype(float) 
secondary_labs_cleaned = secondary_labs_cleaned.dropna(subset=result_col)

# sanity checks for ND and NaNs
display(secondary_labs_cleaned[result_col][secondary_labs_cleaned[result_col].apply(lambda x: not_startswithdigit(x))].unique())
display(secondary_labs_cleaned[result_col].isna().any())

# # tests need to be preprocessed according to Result Values seen for each tests
# # WHAT UNIT IS THIS? LIKELY SAME AS NUHS, use NUHS threshold
# for test in secondary_thresholds.keys():
#     print(test)
#     temp = df.loc[df[test_name_col].isin([test])]
#     display(temp[result_col][temp[result_col].apply(lambda x: not_startswithdigit(x))].unique())
#     # describe
#     display(temp[result_col][temp[result_col].apply(lambda x: not not_startswithdigit(x))].astype(float).describe())
#     print("\n\n")

# patient count checker
# temp = df.loc[df[test_name_col].isin(["TSH RECEPTOR ANTIBODY", "TSH RECEPTOR ANTIBODIES (TRAb)"])]
# temp[temp[result_col].isin(["< 0.90", "< 1.10"])]["Patient ID"].unique().size

secondary_labs_cleaned.loc[secondary_labs_cleaned[test_name_col]=="ABC"]

### 4.3 Thresholding

def threshold_labs(row):
    thresholder = secondary_thresholds.get(row[test_name_col])
    if thresholder:
        return thresholder(row[result_col])
    else:
        return False
    # else:
    #     raise ValueError(f"{row[test_name_col]} not in secondary_thresholds.")

secondary_labs_final = secondary_labs_cleaned.loc[secondary_labs_cleaned.progress_apply(lambda row: threshold_labs(row), axis=1)]

secondary_labs_final["Patient ID"].unique().size

secondary_labs_final.info()

## Filtering for secondary labs based on index date

def check_secondary_labs(
    row, 
    date_col,
    ref_df,
):
    pid = row["Patient ID"]
    index_date = row["Index Date"]

    ref_df[date_col] = pd.to_datetime(ref_df[date_col])
    pid_secondary_labs = ref_df[ref_df["Patient ID"] == pid]
    
    ## take only those records which are between 365 days and 42 days of index date
    pid_secondary_labs = pid_secondary_labs[
        (index_date - pid_secondary_labs[date_col] >= pd.Timedelta(days=0)) & 
        (index_date - pid_secondary_labs[date_col] <= pd.Timedelta(days=365))
    ]
    ## if not empty apply correction else do not
    if (pid_secondary_labs.empty):
        return False
    else:
        return True

tqdm.pandas()
ldlc_valid_18_nontg_index = ldlc_valid_18_nontg_index.reset_index()


date_col = "Specimen Received Date"
with_secondary_labs = ldlc_valid_18_nontg_index[
    ldlc_valid_18_nontg_index.progress_apply(
        check_secondary_labs, args=(date_col, secondary_labs_final), axis=1)
]

with_secondary_labs.info()

with_secondary_labs["Patient ID"].unique().size

with_secondary_labs.to_csv(
    os.path.join(PROJECT_ROOT, "results", "with_secondary_labs.csv"), index=False,
)

ldlc_valid_18_nontg_index_nosecondarylab = ldlc_valid_18_nontg_index.loc[~ldlc_valid_18_nontg_index["Patient ID"].isin(with_secondary_labs["Patient ID"].unique().tolist())]
ldlc_valid_18_nontg_index_nosecondarylab["Patient ID"].unique().size

ldlc_valid_18_nontg_index_nosecondarylab.info()

ldlc_valid_18_nontg_index_nosecondarylab.to_csv(
    os.path.join(PROJECT_ROOT, "results", "ldlc_valid_18_nontg_index_nosecondarylab.csv"), index=False,
)

## End
