==> notebooks/05_diagnosis_filtering.ipynb <==
# Diagnosis Filtering

This notebook covers the creation of references for the following Stage 1 filtering criteria:
1. `Diagnosis` `List` Those with secondary causes within 1 year (+/-) of Index Date

%load_ext autoreload
%autoreload 2

import os, sys
from dotenv import load_dotenv
load_dotenv("../.env")  # take environment variables
PROJECT_ROOT = os.environ.get("PROJECT_ROOT")
sys.path.append(PROJECT_ROOT)

import numpy as np
import pandas as pd
from glob import glob
from tqdm import tqdm
from datetime import datetime

import utils.PATHS as PATHS
import utils.utils as utils
# import utils.emr_utils as emr_utils
# import utils.load_utils as load_utils

dia_fp_list = glob(os.path.join(PATHS.DIAGNOSIS, "*.csv"))

# general labs dataframe
df_list = [pd.read_csv(path, low_memory=False) for path in tqdm(dia_fp_list)]
dia_df = pd.concat(df_list, ignore_index=True)

dia_df.columns

prl_fp_list = glob(os.path.join(PATHS.PROBLEM_LIST, "*.csv"))

df_list = []
for path in tqdm(prl_fp_list):
    temp = pd.read_csv(path, low_memory=False)

    # address div block at df tail
    subset = temp.columns.tolist()
    subset.remove("Visit Date")
    temp = temp.dropna(subset=subset, how="all")

    df_list.append(temp)
    
prl_df = pd.concat(df_list, ignore_index=True)

prl_df.columns

# load ldlc for index date reference
path = os.path.join(PROJECT_ROOT, "results", "ldlc_valid_18_nontg_index_nosecondarylab.csv")
ldlc_valid_18_nontg_index_nosecondarylab = pd.read_csv(path)

# prep datetime cols
ldlc_valid_18_nontg_index_nosecondarylab["Index Date"] = pd.to_datetime(
    ldlc_valid_18_nontg_index_nosecondarylab["Index Date"]
)

def check_secondary_diagnosis(
    row, 
    date_col,
    code_col,
    ref_df,
):
    pid = row["Patient ID"]
    index_date = row["Index Date"]

    ref_df[date_col] = pd.to_datetime(ref_df[date_col])
    pid_dia = ref_df[ref_df["Patient ID"] == pid]
    
    ## take only those records which are between 365 days and 42 days of index date
    pid_dia = pid_dia[
        (index_date - pid_dia[date_col] >= pd.Timedelta(days=0)) & 
        (index_date - pid_dia[date_col] <= pd.Timedelta(days=365))
    ]
    ## if not empty apply correction else do not
    if (pid_dia.empty):
        return False
    return True
    # else:
    #     # TO UDPATE: Break apart snomed search and icd10 search to enable prefix matching in icd10
    #     has_secondary = (
    #         pid_dia[code_col]
    #         .isin(utils.secondary_dia_icd10 + utils.secondary_dia_snomed)
    #         .any()
    #     )
    #     return has_secondary

## SNOMED-coded Problem List : Using Codes

tqdm.pandas()
date_col = "Visit Date"
code_col = "Problem Code (Coded)"
ref_df = prl_df[prl_df[code_col].isin(utils.secondary_dia_icd10 + utils.secondary_dia_snomed)]

with_secondary_prl = ldlc_valid_18_nontg_index_nosecondarylab[
    ldlc_valid_18_nontg_index_nosecondarylab.progress_apply(
        check_secondary_diagnosis, args=(date_col, code_col, ref_df), axis=1)
]

with_secondary_prl["Patient ID"].unique().size

## ICD10-coded Diagnosis: Using Codes

date_col = "Admit/Visit Date"
code_col = "Diagnosis Code (ICD10)"
ref_df = dia_df[dia_df[code_col].isin(utils.secondary_dia_icd10 + utils.secondary_dia_snomed)]

with_secondary_dia = ldlc_valid_18_nontg_index_nosecondarylab.loc[
    ldlc_valid_18_nontg_index_nosecondarylab.progress_apply(
        check_secondary_diagnosis, args=(date_col, code_col, ref_df), axis=1)
]

with_secondary_dia["Patient ID"].unique().size

### SNOMED: Using keywords

def check_secondary_diagnosis(
    row, 
    date_col,
    desc_col,
    ref_df,
):
    pid = row["Patient ID"]
    index_date = row["Index Date"]

    ref_df[date_col] = pd.to_datetime(ref_df[date_col])
    pid_dia = ref_df[ref_df["Patient ID"] == pid]
    
    ## take only those records which are between 365 days and 42 days of index date
    pid_dia = pid_dia[
        (index_date - pid_dia[date_col] >= pd.Timedelta(days=0)) & 
        (index_date - pid_dia[date_col] <= pd.Timedelta(days=365))
    ]
    ## if not empty apply correction else do not
    if (pid_dia.empty):
        return False
    return True
    
    # else:
    #     has_secondary = (
    #         pid_dia[desc_col]
    #         .str.contains('|'.join(utils.secondary_pregnancy_terms), case=False)
    #         .any()
    #     )
    #     return has_secondary

date_col = "Visit Date"
desc_col = "Problem Desc (Coded)"
ref_df = prl_df[prl_df[desc_col].str.contains('|'.join(utils.secondary_pregnancy_terms), case=False)]

with_secondary_prl_keywords = ldlc_valid_18_nontg_index_nosecondarylab.loc[
    ldlc_valid_18_nontg_index_nosecondarylab.progress_apply(
        check_secondary_diagnosis, args=(date_col, desc_col, ref_df), axis=1)
]

with_secondary_prl_keywords["Patient ID"].unique().size

### ICD10: Using keywords

date_col = "Admit/Visit Date"
desc_col = "Diagnosis Description (ICD10)"
ref_df = dia_df[dia_df[desc_col].str.contains('|'.join(utils.secondary_pregnancy_terms), case=False)]
with_secondary_dia_keywords = ldlc_valid_18_nontg_index_nosecondarylab.loc[
    ldlc_valid_18_nontg_index_nosecondarylab.progress_apply(
        check_secondary_diagnosis, args=(date_col, desc_col, ref_df), axis=1)
]

with_secondary_dia_keywords["Patient ID"].unique().size

with_secondary_keywords_plist = with_secondary_prl_keywords["Patient ID"].unique().tolist() + with_secondary_dia_keywords["Patient ID"].unique().tolist()

len(with_secondary_keywords_plist)

## Post-secondary filtering

ldlc_valid_18_nontg_index_nosecondary_final = ldlc_valid_18_nontg_index_nosecondarylab.loc[
    ~ldlc_valid_18_nontg_index_nosecondarylab["Patient ID"].isin(with_secondary_keywords_plist)
]

ldlc_valid_18_nontg_index_nosecondary_final["Patient ID"].unique().size

ldlc_valid_18_nontg_index_nosecondary_final.to_csv(
    os.path.join(PROJECT_ROOT, "results", "ldlc_valid_18_nontg_index_nosecondary_final.csv"), index=False,
)

## End
